% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xegaGpFindExample.R
\name{findCrossoverExample}
\alias{findCrossoverExample}
\title{Prints a random example of crossover for a crossover method given 
a random number seed.}
\usage{
findCrossoverExample(FUN, s, verbose = TRUE)
}
\arguments{
\item{FUN}{String. Specification of crossover method.}

\item{s}{Integer. Seed of random number generator.}

\item{verbose}{Boolean. 
If \code{TRUE} (default), print example to console.}
}
\value{
No return.
}
\description{
The purpose of this function is to support the search 
             for examples for generating unit tests for crossover 
             functions whose behavior depends on random numbers.
}
\examples{
findCrossoverExample(FUN="AllCross2Gene", s=2)
}
\concept{Testing}
