% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xegaGpDecode.R
\name{xegaGpDecodeGene}
\alias{xegaGpDecodeGene}
\title{Decode a derivation tree.}
\usage{
xegaGpDecodeGene(gene, lF)
}
\arguments{
\item{gene}{Derivation tree.}

\item{lF}{Local configuration of the genetic algorithm.}
}
\value{
Decoded gene.
}
\description{
\code{xegaGpDecodeGene()} decodes a derivation tree.
}
\details{
The recursive algorithm for the decoder is imported 
         from the package \code{xegaDerivationTrees}.
}
\examples{
gene<-xegaGpInitGene(lFxegaGpGene)
xegaGpDecodeGene(gene, lFxegaGpGene)

}
\concept{Decoder}
