% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xegaGplF.R
\docType{data}
\name{lFxegaGpGene}
\alias{lFxegaGpGene}
\title{Generate local functions and objects.}
\format{
An object of class \code{list} of length 25.
}
\usage{
lFxegaGpGene
}
\description{
\code{lFxegaPermGene} is a list of functions 
which contains a definition of all local objects 
required for the use of genetic operators with the 
We refer to this object as local configuration.
}
\keyword{datasets}
